Reorganize code to abide to new import rule
