+++ packages/liveblocks-core/src/crdts/__tests__/liveblocks-helpers.test.ts
@@ -2,10 +2,10 @@ import {
   findNonSerializableValue,
   getTreesDiffOperations,
 } from "../liveblocks-helpers";
-import { OpCode } from "../protocol/Op";
-import { CrdtType } from "../protocol/SerializedCrdt";
-import type { NodeMap } from "../types/NodeMap";
-import { FIRST_POSITION, SECOND_POSITION } from "./_utils";
+import { OpCode } from "../../protocol/Op";
+import { CrdtType } from "../../protocol/SerializedCrdt";
+import type { NodeMap } from "../../types/NodeMap";
+import { FIRST_POSITION, SECOND_POSITION } from "../../__tests__/_utils";
 
 describe("getTreesDiffOperations", () => {
   test("new liveList Register item", () => {
