+++ packages/liveblocks-core/src/crdts/liveblocks-helpers.ts
@@ -1,18 +1,18 @@
-import type { ManagedPool } from "./crdts/AbstractCrdt";
-import { type LiveListUpdates, LiveList } from "./crdts/LiveList";
-import { type LiveMapUpdates, LiveMap } from "./crdts/LiveMap";
-import { type LiveObjectUpdates, LiveObject } from "./crdts/LiveObject";
-import { LiveRegister } from "./crdts/LiveRegister";
-import type { LiveNode, LiveStructure, Lson, LsonObject } from "./crdts/Lson";
-import type { StorageUpdate } from "./crdts/StorageUpdates";
-import { assertNever, nn } from "./lib/assert";
-import type { Json } from "./lib/Json";
-import { entries, isPlainObject } from "./lib/utils";
-import type { CreateOp, Op } from "./protocol/Op";
-import { OpCode } from "./protocol/Op";
-import type { IdTuple, SerializedCrdt } from "./protocol/SerializedCrdt";
-import { CrdtType } from "./protocol/SerializedCrdt";
-import type { NodeMap, ParentToChildNodeMap } from "./types/NodeMap";
+import type { ManagedPool } from "./AbstractCrdt";
+import { type LiveListUpdates, LiveList } from "./LiveList";
+import { type LiveMapUpdates, LiveMap } from "./LiveMap";
+import { type LiveObjectUpdates, LiveObject } from "./LiveObject";
+import { LiveRegister } from "./LiveRegister";
+import type { LiveNode, LiveStructure, Lson, LsonObject } from "./Lson";
+import type { StorageUpdate } from "./StorageUpdates";
+import { assertNever, nn } from "../lib/assert";
+import type { Json } from "../lib/Json";
+import { entries, isPlainObject } from "../lib/utils";
+import type { CreateOp, Op } from "../protocol/Op";
+import { OpCode } from "../protocol/Op";
+import type { IdTuple, SerializedCrdt } from "../protocol/SerializedCrdt";
+import { CrdtType } from "../protocol/SerializedCrdt";
+import type { NodeMap, ParentToChildNodeMap } from "../types/NodeMap";
 
 export function creationOpToLiveNode(op: CreateOp): LiveNode {
   return lsonToLiveNode(creationOpToLson(op));
